web_sqlite3/requirements.txt
aiosqlite>=0.19.0
pyyaml>=6.0
